Add symlink at old path to blueprint
On the old drupal dissco site, the blueprint was found at https://www.dissco-uk.org/sites/default/files/2022-07/Harnessing%20the%20Power%20of%20Natural%20Science%20Collections.pdf which is also used in a bitly link we shared. We can't update the link without paying sweet cash, so this is just a simple redirect to make sure that path is still available to users.